Move scheduler_plugin.go to plugins package
