+++ pkg/plugin/scheduler_plugin.go
@@ -16,7 +16,7 @@
  limitations under the License.
 */
 
-package schedulerplugin
+package plugin
 
 import (
 	"context"
